% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamanagement.R
\name{comp_int_q_WNN}
\alias{comp_int_q_WNN}
\title{Compute Intermediate Quantities for Basis Function Evaluation (when a weighted nearest neighbours interpolation is done)}
\usage{
comp_int_q_WNN(
  normalizedData,
  predictorNames,
  responseName,
  nIntegral = 101,
  nDiscret = 51
)
}
\arguments{
\item{normalizedData}{A dataframe containing the normalized dataset.}

\item{predictorNames}{A character vector specifying the names of the predictor variables.}

\item{responseName}{A character string specifying the name of the response variable.}

\item{nIntegral}{Number of points used to approximate the integral.}

\item{nDiscret}{Number of points used to discretize the predictors' domain.}
}
\value{
A list containing intermediate quantities for basis function evaluation.
}
\description{
This function takes normalized data, predictor names, response name,
and computes intermediate quantities useful for later evaluation of basis functions.
}
\examples{
data <- data.frame(x1 = c(0.1, 0.5, 0.9), x2 = c(0.2, 0.6, 1.0), y = c(0.3, 0.7, 1.0))
normalized_data <- normalize_data(data, c("x1", "x2"), "y")
intermediate_quantities <- comp_int_q_WNN(normalized_data, c("x1", "x2"), "y")

}
