% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{crossdist}
\alias{crossdist}
\title{Computes  the Euclidean distance between rows of two matrices}
\usage{
crossdist(x, y)
}
\arguments{
\item{x}{First matrix}

\item{y}{Second matrix}
}
\value{
Euclidean distance between rows of \code{x} and \code{y}
}
\description{
Computes  the Euclidean distance between rows of two matrices
}
