% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisFunInitAndCompute.R
\name{heuristic_find_variance}
\alias{heuristic_find_variance}
\title{Heuristic function to find the range of values of the unconditioned SLGP.}
\usage{
heuristic_find_variance(
  parameters,
  nsimu = 1000,
  grid_size = 101,
  plot = FALSE
)
}
\arguments{
\item{parameters}{A list containing outputs of initialize_basisfun.}

\item{nsimu}{An integer giving the number of unconditional simulations to use.}

\item{grid_size}{An integer giving the number of nodes in each dimension of the grid to consider.}

\item{plot}{A boolean indicating whether a ngraphical output is produced.}
}
\value{
A matrix with the evaluated basis functions.
}
\description{
A Heuristic function to find the empirical range of the unconditioned SLGP
}
