% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{negloglike}
\alias{negloglike}
\title{Computes the negative log likelihood}
\usage{
negloglike(epsilon, y)
}
\arguments{
\item{y}{Second matrix}

\item{x}{First matrix}
}
\value{
something
}
\description{
Computes the negative log likelihood
}
