% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slgp.R
\name{newSLGP}
\alias{newSLGP}
\title{Create a new SLGP object without performing data-based estimation}
\usage{
newSLGP(
  basisFunctionsUsed,
  w = NULL,
  p,
  hyperparams = NULL,
  sigmaEstimationMethod = "none",
  opts = list(),
  response_range = NULL,
  predictor_ranges = NULL
)
}
\arguments{
\item{basisFunctionsUsed}{String specifying the basis functions ("poly" and/or "sines").}

\item{w}{Optional vector of weights.}

\item{p}{Number of basis functions.}

\item{hyperparams}{Optional hyper-parameter values.}

\item{sigmaEstimationMethod}{Method for estimating sigma ("none" or "heuristic").}

\item{opts}{Optional list of extra parameters.}

\item{response_range}{A vector specifying the expected range of the response variable.}

\item{predictor_ranges}{A list specifying the expected range for each predictor variable.}

\item{method}{The method to be used ("MCMC", "MAP", "Laplace").}
}
\value{
An instance of the SLGP class.
}
\description{
Create a new SLGP object without performing data-based estimation
}
