% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictAndSimulate.R
\name{predictSLGP_moments}
\alias{predictSLGP_moments}
\title{Perform prediction at candidate points of the (un)centered moment(s) in a SLGP model.}
\usage{
predictSLGP_moments(
  SLGPmodel,
  newNodes,
  power,
  centered = FALSE,
  interpolateBasisFun = "WNN",
  nIntegral = 101,
  nDiscret = 101
)
}
\arguments{
\item{SLGPmodel}{An object of class SLGP.}

\item{newNodes}{A data frame containing the new points at which we want the SLGP(s) evaluated}

\item{power}{Scalar or vector, specifying the moments to be predicted.}

\item{centered}{Boolean, specifying if the moments to be predicted are centered (TRUE) or not (FALSE).}

\item{interpolateBasisFun}{String specifying whether the basis functions are evaluated on all points ("nothing"), on the closest neighbour of a regular grid ("NN" - default) or with a weighted inverse distance to the closest neighbours ("WID").}

\item{nIntegral}{Number of points used to approximate the integral.}

\item{nDiscret}{Integer, optional, discretization step used if "interpolateBasisFun" is "NN" or "WNN".}
}
\value{
A list containing the results of the SLGP regression.
}
\description{
Perform prediction at candidate points of the (un)centered moment(s) in a SLGP model.
}
