% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{rosenblatt_transform_multivarStudent}
\alias{rosenblatt_transform_multivarStudent}
\title{Rosenblatt transform to multivariate Student distribution}
\usage{
rosenblatt_transform_multivarStudent(x, dimension, MatParam = 5/2)
}
\arguments{
\item{x}{A matrix (or vector) of samples in [0, 1]^d to transform.}

\item{dimension}{Integer. The dimension of the input space.}

\item{MatParam}{Numeric. The Matérn kernel smoothness parameter (default = 5/2).}
}
\value{
A matrix with transformed coordinates following a multivariate Student distribution.
}
\description{
Auxiliary function that maps uniform samples in [0, 1]^d to samples from the spectral density
of a Matérn kernel (i.e., a multivariate Student distribution).
}
\keyword{internal}
