% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLGPclass.R
\docType{class}
\name{SLGP-class}
\alias{SLGP-class}
\alias{SLGP}
\title{Spatial Logistic Gaussian Process Class}
\description{
Spatial Logistic Gaussian Process Class
}
\section{Slots}{

\describe{
\item{\code{formula}}{Formula specifying the covariates.}

\item{\code{data}}{A data frame containing the data to train the SLGP.}

\item{\code{responseName}}{A character, specifying the name of the response.}

\item{\code{covariateName}}{A character vector, specifying the names of the covariates}

\item{\code{responseRange}}{A vector with the response upper range and lower range.}

\item{\code{predictorsRange}}{A list containing the vector 'predictorsLower' specifying the covariate's lower range
and the vector 'predictorsUpper' specifying the covariate's upper range}

\item{\code{method}}{The method used to train the SLGP among {"MCMC", "MAP", "Laplace", "none"}.}

\item{\code{p}}{Number of basis functions.}

\item{\code{basisFunctionsUsed}}{String specifying the basis functions ("inducing points", "RFF", "Discrete FF", "filling FF", "custom cosines").}

\item{\code{opts_BasisFun}}{List of extra parameters for the basis functions.}

\item{\code{coefficients}}{Matrix of epsilon's values for the finite-rank GP:
\eqn{ Z(x,t) = \sum_{i=1}^p \epsilon_i f_i(x, t) }}

\item{\code{hyperparams}}{Hyper-parameter values.It should be a list with a numeric 'sigma' and a vector 'lengthscale'.}
}}

