% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLGP-package.R
\docType{package}
\name{SLGP-package}
\alias{SLGP-package}
\alias{SLGP}
\title{SLGP: A package for performing sample-based estimation of spatially dependent probability distributions.}
\section{SLGP functions}{

The SLGP functions ...
}

