% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisFunInitAndCompute.R
\name{evaluate_basis_functions}
\alias{evaluate_basis_functions}
\title{Evaluate a Basis of Functions at Given Locations}
\usage{
evaluate_basis_functions(parameters, X, lengthscale)
}
\arguments{
\item{parameters}{A list containing outputs of initialize_basisfun.}

\item{X}{A design matrix containing locations where we want to evaluate the function.}

\item{lengthscale}{Numeric vector containing the lengthscales to use for the kernel.}
}
\value{
A matrix with the evaluated basis functions.
}
\description{
Evaluate a basis of functions at given locations.
}
