% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisFunInitAndCompute.R
\name{initialize_basisfun_discreteFF}
\alias{initialize_basisfun_discreteFF}
\title{Initialize discrete Fourier features}
\usage{
initialize_basisfun_discreteFF(dimension, maxOrdert, maxOrderx)
}
\arguments{
\item{dimension}{Integer. Input (\eqn{[\mathbf{x},\,t]}{[x, t]}) dimension.}

\item{maxOrdert}{Integer. Maximum frequency in t.}

\item{maxOrderx}{Integer. Maximum frequency in each x.}
}
\value{
List with frequency, offset, and coefficient parameters.
}
\description{
Generates basis using discrete cosine/sine terms for each input dimension.
}
\keyword{internal}
