% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisFunInitAndCompute.R
\name{initialize_basisfun_fillingRFF}
\alias{initialize_basisfun_fillingRFF}
\title{Initialize parameters for basis functions based on space-filling Random Fourier Features.}
\usage{
initialize_basisfun_fillingRFF(
  dimension,
  nFreq,
  MatParam = 5/2,
  lengthscale,
  seed = 0
)
}
\arguments{
\item{dimension}{Numeric. The dimension of the index
\eqn{[\mathbf{x},\,t]}{[x, t]}.}

\item{nFreq}{Numeric. Number of frequencies to sample.}

\item{MatParam}{Numeric, specifying the parameter of the Matérn kernel considered (default = 5/2).}

\item{lengthscale}{Numeric vector containing the lengthscales to use for the kernel.}
}
\value{
List. A list containing the initialized parameters necessary for evaluating the specified basis function.
}
\description{
This function initializes parameters for basis functions based on space-filling Random Fourier Features (for Matérn kernels only).
}
\examples{
initialize_basisfun_fillingRFF(dimension = 2, nFreq=10, MatParam = 5/2, lengthscale=c(0.1, 0.1))

}
