% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisFunInitAndCompute.R
\name{initialize_basisfun_inducingpt}
\alias{initialize_basisfun_inducingpt}
\title{Initialize parameters for Inducing points based functions.}
\usage{
initialize_basisfun_inducingpt(
  dimension,
  kernel = "Mat52",
  lengthscale,
  pointscoord = NULL,
  numberPoints = NULL
)
}
\arguments{
\item{dimension}{Numeric. The dimension of the index
\eqn{[\mathbf{x},\,t]}{[x, t]}.}

\item{kernel}{Character, specifying the kernel to be consider among "Gaussian", "Exp", "Mat32" and "Mat52" (default "Mat52").}

\item{lengthscale}{Numeric vector containing the lengthscales to use for the kernel.}

\item{pointscoord}{Optional matrix with the coordinates of the inducing points. If none is provided, we sample them uniformly in the unit hypercube.}

\item{numberPoints}{Optional numerical value specifying the number of inducing points to sample (ignored if \code{pointscoord} is specified)}
}
\value{
List. A list containing the upper triangular factor of the Cholesky decomposition of the kernel matrix as well as its inverse.
}
\description{
This function initializes parameters for basis functions based on inducing points.
}
\examples{
initialize_basisfun_inducingpt(dimension = 2, lengthscale=c(0.1, 0.1), numberPoints=5)

}
