% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamanagement.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize Data to the Range \link{0, 1}}
\usage{
normalize_data(
  data,
  predictorNames,
  responseName,
  predictorsUpper = NULL,
  predictorsLower = NULL,
  responseRange = NULL
)
}
\arguments{
\item{data}{A dataframe containing the dataset.}

\item{predictorNames}{A character vector specifying the names of the predictor variables.}

\item{responseName}{A character string specifying the name of the response variable.}

\item{predictorsUpper}{A numeric vector representing the upper range for the predictors (optional).}

\item{predictorsLower}{A numeric vector representing the lower range for the predictors (optional).}

\item{responseRange}{A numeric vector representing the upper and lower range for the response (optional).}
}
\value{
A dataframe with the normalized values.
}
\description{
This function takes a dataframe, a vector of predictor names, a response name, and optional range information,
and normalizes the data to the range \link{0, 1}. If range information is not provided, the observed
values in the data are used to determine the range.
}
\examples{
data <- data.frame(x1 = c(1, 2, 3), x2 = c(4, 5, 6), y = c(10, 20, 30))
normalized_data <- normalize_data(data, c("x1", "x2"), "y")

}
