% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slgp.R
\name{retrainSLGP}
\alias{retrainSLGP}
\title{Retrain a SLGP by changing the data and/or method.}
\usage{
retrainSLGP(
  SLGPmodel,
  newdata = NULL,
  epsilonStart = NULL,
  method,
  interpolateBasisFun = "NN",
  nIntegral = 51,
  nDiscret = 51,
  hyperparams = NULL,
  sigmaEstimationMethod = "none",
  seed = NULL,
  opts = list()
)
}
\arguments{
\item{SLGPmodel}{A SLGP.}

\item{newdata}{An optional data frame containing the variables in the formula.}

\item{epsilonStart}{An optional numeric vector, the starting weights in the finite-rank GP:
\eqn{ Z(x,t) = \sum_{i=1}^p \epsilon_i f_i(x, t) }}

\item{method}{The method to be used among {"MCMC", "MAP", "Laplace"}.}

\item{interpolateBasisFun}{String specifying whether the basis functions are evaluated on all points ("nothing"), on the closest neighbour of a regular grid ("NN" - default) or with a weighted inverse distance to the closest neighbours ("WID").}

\item{nIntegral}{Number of points used to approximate the integral.}

\item{nDiscret}{Integer, optional, discretization step used if "interpolateBasisFun" is "NN" or "WNN".}

\item{hyperparams}{Optional hyper-parameter values. It should be a list with sigma and a vector for lengthscale.}

\item{sigmaEstimationMethod}{Method for estimating sigma2 ("none" (default) or "heuristic").}

\item{opts}{Optional list of extra parameters, typically for the MCMC or optimisation.}
}
\value{
A list containing the results of the SLGP regression.
}
\description{
Creates a \code{slgp} object and performs the training using either a Bayesian MCMC estimation, a MAP estimation or a Laplace approximation (i.e. MAP + Laplace).
}
