% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictAndSimulate.R
\name{sampleSLGP}
\alias{sampleSLGP}
\title{Draw new samples from a SLGP model}
\usage{
sampleSLGP(
  SLGPmodel,
  newX,
  n,
  interpolateBasisFun = "NN",
  nIntegral = 51,
  nDiscret = 51,
  seed = NULL
)
}
\arguments{
\item{SLGPmodel}{An object of class SLGP.}

\item{newX}{A data frame containing the new points at which we want draws from a SLGP}

\item{n}{An integer, (or vector of integers with length matching the number of rows in newPredictors) specifying the number of samples to be drawn.}

\item{interpolateBasisFun}{String specifying whether the basis functions are evaluated on all points ("nothing"), on the closest neighbour of a regular grid ("NN" - default) or with a weighted inverse distance to the closest neighbours ("WID").}

\item{nIntegral}{Number of points used to approximate the integral.}

\item{nDiscret}{Integer, optional, discretization step used if "interpolateBasisFun" is "NN" or "WNN".}

\item{seed}{Optional, to specify a seed}
}
\value{
A list containing the results of the SLGP regression.
}
\description{
Draw new samples from a SLGP model
}
