% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLGP-package.R
\name{SLGP-package}
\alias{SLGP-package}
\alias{SLGP}
\title{SLGP: A package for performing sample-based estimation of spatially dependent probability distributions.}
\description{
SLGP: A package for performing sample-based estimation of spatially dependent probability distributions.
}
\section{SLGP functions}{

The SLGP functions ...
}

\references{
Gautier, Athénaïs (2023). "Modelling and Predicting Distribution-Valued Fields with Applications to Inversion Under Uncertainty." Thesis, Universität Bern, Bern.
\link{https://boristheses.unibe.ch/4377/}
}
