% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slgp.R
\name{slgp}
\alias{slgp}
\title{Perform SLGP estimation using method of choice.}
\usage{
slgp(
  formula,
  data,
  epsilonStart = NULL,
  method,
  basisFunctionsUsed,
  interpolateBasisFun = "NN",
  nIntegral = 51,
  nDiscret = 51,
  hyperparams = NULL,
  predictorsUpper = NULL,
  predictorsLower = NULL,
  responseRange = NULL,
  sigmaEstimationMethod = "none",
  seed = NULL,
  opts_BasisFun = list(),
  BasisFunParam = NULL,
  opts = list()
)
}
\arguments{
\item{formula}{A formula specifying the model.}

\item{data}{A data frame containing the variables in the formula.}

\item{epsilonStart}{An optional numeric vector, the starting weights in the finite-rank GP:
\eqn{ Z(x,t) = \sum_{i=1}^p \epsilon_i f_i(x, t) }}

\item{method}{The method to be used among {"none", "MCMC", "MAP", "Laplace"}.}

\item{basisFunctionsUsed}{String specifying the basis functions ("inducing points", "RFF", "Discrete FF", "filling FF", "custom cosines").}

\item{interpolateBasisFun}{String specifying whether the basis functions are evaluated on all points ("nothing"), on the closest neighbour of a regular grid ("NN" - default) or with a weighted inverse distance to the closest neighbours ("WNN").}

\item{nIntegral}{Number of points used to approximate the integral.}

\item{nDiscret}{Integer, optional, discretization step used if "interpolateBasisFun" is "NN" or "WNN".}

\item{hyperparams}{Optional hyper-parameter values. It should be a list with sigma and a vector for lengthscale.}

\item{predictorsUpper}{An optional vector with the response upper range and lower range.}

\item{predictorsLower}{An optional vector with the response upper range and lower range.}

\item{responseRange}{An optional vector with the response upper range and lower range.}

\item{sigmaEstimationMethod}{Method for estimating sigma2 ("none" (default) or "heuristic").}

\item{opts_BasisFun}{List of extra parameters for the basis functions.}

\item{BasisFunParam}{List to specify the basis functions}

\item{opts}{Optional list of extra parameters, typically for the MCMC or optimisation.}
}
\value{
A list containing the results of the SLGP regression.
}
\description{
Creates a \code{slgp} object and performs the training using either a Bayesian MCMC estimation, a MAP estimation or a Laplace approximation (i.e. MAP + Laplace).
}
\examples{
\dontrun{
}
}
\references{
Gautier, Athénaïs (2023). "Modelling and Predicting Distribution-Valued Fields with Applications to Inversion Under Uncertainty." Thesis, Universität Bern, Bern.
\link{https://boristheses.unibe.ch/4377/}
}
